Split storm-dft.h into analysis.h, io.h and transformation.h
