+++ src/test/storm-dft/api/DftSmtTest.cpp
@@ -2,6 +2,9 @@
 #include "test/storm_gtest.h"
 
 #include "storm-dft/api/storm-dft.h"
+#include "storm-dft/modelchecker/DFTASFChecker.h"
+#include "storm-dft/utility/FDEPConflictFinder.h"
+#include "storm-dft/utility/FailureBoundFinder.h"
 
 namespace {
 class DftSmt : public ::testing::Test {
